+++ management/model_create_application_request.go
@@ -15,8 +15,16 @@ import (
 
 // CreateApplicationRequest - struct for CreateApplicationRequest
 type CreateApplicationRequest struct {
-	ApplicationOIDC *ApplicationOIDC
-	ApplicationSAML *ApplicationSAML
+	ApplicationExternalLink *ApplicationExternalLink
+	ApplicationOIDC         *ApplicationOIDC
+	ApplicationSAML         *ApplicationSAML
+}
+
+// ApplicationExternalLinkAsCreateApplicationRequest is a convenience function that returns ApplicationExternalLink wrapped in CreateApplicationRequest
+func ApplicationExternalLinkAsCreateApplicationRequest(v *ApplicationExternalLink) CreateApplicationRequest {
+	return CreateApplicationRequest{
+		ApplicationExternalLink: v,
+	}
 }
 
 // ApplicationOIDCAsCreateApplicationRequest is a convenience function that returns ApplicationOIDC wrapped in CreateApplicationRequest
@@ -44,6 +52,7 @@ func (dst *CreateApplicationRequest) UnmarshalJSON(data []byte) error {
 
 	dst.ApplicationOIDC = nil
 	dst.ApplicationSAML = nil
+	dst.ApplicationExternalLink = nil
 
 	switch common.GetProtocol() {
 	case ENUMAPPLICATIONPROTOCOL_OPENID_CONNECT:
@@ -54,6 +63,10 @@ func (dst *CreateApplicationRequest) UnmarshalJSON(data []byte) error {
 		if err := json.Unmarshal(data, &dst.ApplicationSAML); err != nil { // simple model
 			return err
 		}
+	case ENUMAPPLICATIONPROTOCOL_EXTERNAL_LINK:
+		if err := json.Unmarshal(data, &dst.ApplicationExternalLink); err != nil { // simple model
+			return err
+		}
 	default:
 		return fmt.Errorf("Data failed to match schemas in oneOf(CreateApplicationRequest)")
 	}
@@ -62,6 +75,10 @@ func (dst *CreateApplicationRequest) UnmarshalJSON(data []byte) error {
 
 // Marshal data from the first non-nil pointers in the struct to JSON
 func (src CreateApplicationRequest) MarshalJSON() ([]byte, error) {
+	if src.ApplicationExternalLink != nil {
+		return json.Marshal(&src.ApplicationExternalLink)
+	}
+
 	if src.ApplicationOIDC != nil {
 		return json.Marshal(&src.ApplicationOIDC)
 	}
@@ -78,6 +95,10 @@ func (obj *CreateApplicationRequest) GetActualInstance() interface{} {
 	if obj == nil {
 		return nil
 	}
+	if obj.ApplicationExternalLink != nil {
+		return obj.ApplicationExternalLink
+	}
+
 	if obj.ApplicationOIDC != nil {
 		return obj.ApplicationOIDC
 	}